+++ R/quarto_ext.R
@@ -253,11 +253,14 @@ build_app_resources <- function(app_json) {
     }
   })
 
-  # Download wasm binaries ready to embed into Quarto deps
-  withr::with_options(
-    list(shinylive.quiet = TRUE),
-    download_wasm_packages(appdir, destdir, package_cache = TRUE, max_filesize = NULL)
-  )
+  wasm_packages <- sys_env_wasm_packages()
+  if (wasm_packages) {
+    # Download wasm binaries ready to embed into Quarto deps
+    withr::with_options(
+      list(shinylive.quiet = TRUE),
+      download_wasm_packages(appdir, destdir, package_cache = TRUE, max_filesize = NULL)
+    )
+  }
 
   # Enumerate R package Wasm binaries and prepare the VFS images as html deps
   webr_dir <- fs::path(destdir, "shinylive", "webr")